Travis (#14)
* first draft of .travis.yml

* add missing sklearn *test only* dependency

* create completion helpers file if not exists

* create completion helpers file if not exists

* create completion helpers file if not exists

* coveralls and badges

* bump introspection version for full test suite passing on python3.8

+++ bourbaki/application/completion/completers.py
@@ -539,10 +539,15 @@ def _install_shell_completion_helpers(completions_helpers_file):
     source = pkgutil.get_data('bourbaki.application.completion', BASH_COMPLETION_HELPERS_FILENAME).strip()
     completions_helpers_file_absolute = os.path.expanduser(completions_helpers_file)
 
-    with open(completions_helpers_file_absolute, "rb+") as outfile:
-        old_source = outfile.read().strip()
-        if source != old_source:
-            print("REINSTALLING BASH COMPLETION HELPERS AT {}".format(completions_helpers_file_absolute), file=sys.stderr)
-            outfile.seek(0)
-            outfile.truncate()
+    if not os.path.exists(completions_helpers_file_absolute):
+        with open(completions_helpers_file_absolute, "wb") as outfile:
+            print("INSTALLING BASH COMPLETION HELPERS AT {}".format(completions_helpers_file_absolute), file=sys.stderr)
             outfile.write(source)
+    else:
+        with open(completions_helpers_file_absolute, "rb+") as outfile:
+            old_source = outfile.read().strip()
+            if source != old_source:
+                print("REINSTALLING BASH COMPLETION HELPERS AT {}".format(completions_helpers_file_absolute), file=sys.stderr)
+                outfile.seek(0)
+                outfile.truncate()
+                outfile.write(source)
